+++ TeamCode/src/main/java/com/acmerobotics/relicrecovery/opmodes/AutoPaths.java
@@ -6,7 +6,7 @@ import com.acmerobotics.library.localization.Vector2d;
 import com.acmerobotics.relicrecovery.configuration.BalancingStone;
 import com.acmerobotics.relicrecovery.configuration.Cryptobox;
 import com.acmerobotics.relicrecovery.path.Path;
-import com.acmerobotics.relicrecovery.path2.PathBuilder;
+import com.acmerobotics.relicrecovery.path.PathBuilder;
 
 import org.firstinspires.ftc.robotcore.external.navigation.RelicRecoveryVuMark;
