+++ TeamCode/src/main/java/com/acmerobotics/relicrecovery/opmodes/test/MoveTest.java
@@ -1,7 +1,7 @@
 package com.acmerobotics.relicrecovery.opmodes.test;
 
 import com.acmerobotics.relicrecovery.opmodes.AutoOpMode;
-import com.acmerobotics.relicrecovery.path2.PathBuilder;
+import com.acmerobotics.relicrecovery.path.PathBuilder;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 
 @Autonomous
