+++ TeamCode/src/main/java/com/acmerobotics/relicrecovery/opmodes/test/StrafeTest.java
@@ -3,7 +3,7 @@ package com.acmerobotics.relicrecovery.opmodes.test;
 import com.acmerobotics.library.dashboard.config.Config;
 import com.acmerobotics.relicrecovery.opmodes.AutoOpMode;
 import com.acmerobotics.relicrecovery.opmodes.AutoPaths;
-import com.acmerobotics.relicrecovery.path2.PathBuilder;
+import com.acmerobotics.relicrecovery.path.PathBuilder;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 
 @Autonomous
