+++ TeamCode/src/main/java/com/acmerobotics/relicrecovery/opmodes/tuner/AxialFFTuner.java
@@ -3,8 +3,8 @@ package com.acmerobotics.relicrecovery.opmodes.tuner;
 import com.acmerobotics.library.dashboard.config.Config;
 import com.acmerobotics.library.localization.Pose2d;
 import com.acmerobotics.library.localization.Vector2d;
-import com.acmerobotics.relicrecovery.path2.PathBuilder;
-import com.acmerobotics.relicrecovery.path2.PathFollower;
+import com.acmerobotics.relicrecovery.path.PathBuilder;
+import com.acmerobotics.relicrecovery.path.PathFollower;
 import com.acmerobotics.relicrecovery.subsystems.MecanumDrive;
 import com.acmerobotics.relicrecovery.subsystems.Robot;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
