+++ TeamCode/src/main/java/com/acmerobotics/relicrecovery/path/PathMotionSegment.java
@@ -1,4 +1,4 @@
-package com.acmerobotics.relicrecovery.path2;
+package com.acmerobotics.relicrecovery.path;
 
 import com.acmerobotics.library.localization.Pose2d;
